+++ src/component/3d/light.cpp
@@ -27,7 +27,7 @@ void CD3DLight::Uninit()
 	std::vector<CD3DLight*> lightList = GetComponents<CD3DLight>(false, true);
 
 	// ÅŒã”ö‚Ìƒ‰ƒCƒg‚ð–³Œø‚É‚·‚é
-	pDevice->LightEnable(lightList.size() - 1, FALSE);
+	pDevice->LightEnable(static_cast<DWORD>(lightList.size() - 1), FALSE);
 }
 
 //=============================================================
